Add helper function
Add function that returns the staking app label based on the
`StakingAppName` type.
