+++ src/components/Modal/DeauthorizeApplicationModal/InititateDeauthorization.tsx
@@ -24,11 +24,7 @@ import shortenAddress from "../../../utils/shortenAddress"
 import TokenBalance from "../../TokenBalance"
 import { StakingAppName } from "../../../store/staking-applications"
 import { useInitiateDeauthorization } from "../../../hooks/staking-applications"
-
-const stakingAppNameToAppLabel: Record<StakingAppName, string> = {
-  tbtc: "tBTC",
-  randomBeacon: "Random Beacon",
-}
+import { getSakingAppLabel } from "../../../utils/getStakingAppLabel"
 
 const InitiateDeauthorization: FC<{
   closeModal: () => void
@@ -50,7 +46,7 @@ const InitiateDeauthorization: FC<{
         <InfoBox variant="modal">
           <H5 mb={4}>
             You're about to initiate the decrease of your{" "}
-            {stakingAppNameToAppLabel[stakingAppName]} authorization.
+            {getSakingAppLabel(stakingAppName)} authorization.
           </H5>
           <BodyLg>
             Initiation and confirmation of deauthorization is a two step action.
