--- src/utils/getStakingAppNameFromAddress.ts
@@ -1,10 +1,0 @@
-import { threshold } from "./getThresholdLib"
-
-const stakingAppAddressToName: { [key: string]: string } = {
-  [threshold.multiAppStaking.ecdsa.address]: "tBTC",
-  [threshold.multiAppStaking.randomBeacon.address]: "Random Beacon",
-}
-
-export const getStakingAppNameFromAddress = (stakingAppAddress: string) => {
-  return stakingAppAddressToName[stakingAppAddress] ?? "App"
-}
